# docs/swagger.yaml
consumes:
- application/json
definitions:
  rest.AuthResp:
    properties:
      jwt_token:
        type: string
      nickname:
        example: phd ***** destroyer
        type: string
      role:
        allOf:
        - $ref: '#/definitions/storage.ParticipantRole'
        example: 1
    type: object
...
  rest.BecomeValidatorRequest:
    properties:
      email_address:
        description: mandatory
        type: string
      language:
        description: mandatory
        type: string
      middlename:
        type: string
      name:
        description: mandatory
        type: string
      orcid:
        type: string
      sciences:
        items:
          type: string
        type: array
      surname:
        description: mandatory
        type: string
    type: object
  rest.ErrorMsg:
    properties:
      error:
        example: null request param
        type: string
    type: object
...
  rest.SuccessMsg:
    properties:
      status:
        example: OK
        type: string
    type: object
...
  rest.UpdateValidatorRequest:
    properties:
      email_address:
        type: string
      language:
        type: string
      middlename:
        type: string
      name:
        type: string
      orcid:
        type: string
      sciences:
        items:
          type: string
        type: array
      surname:
        type: string
    type: object
  rest.WorkReviewRequest:
    properties:
      review:
        $ref: '#/definitions/storage.WorkReview'
    type: object
...
  storage.Participant:
    properties:
      language:
        description: '''ru'', ''en'''
        type: string
      nickname:
        type: string
      role:
        $ref: '#/definitions/storage.ParticipantRole'
      web3Address:
        type: string
    type: object
  storage.ParticipantRole:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-comments:
      AdminRole: 5 Admin
      AdvisorRole: 3 Advisor
      AuthorRole: 2 Author
      GuestRole: 0 Guest
      ReaderRole: 1 Reader
      ValidatorRole: 4 Validator
    x-enum-varnames:
    - GuestRole
    - ReaderRole
    - AuthorRole
    - AdvisorRole
    - ValidatorRole
    - AdminRole
  storage.Validator:
    properties:
      diploma_id:
        description: referrenceKey
        type: string
      email_address:
        type: string
      language:
        type: string
      middlename:
        type: string
      name:
        type: string
      orcid:
        type: string
      sciences:
        items:
          type: string
        type: array
      surname:
        type: string
    type: object
  storage.ValidatorResponse:
    properties:
      basic_info:
        $ref: '#/definitions/storage.Participant'
      validator_info:
        $ref: '#/definitions/storage.Validator'
    type: object
...
  storage.WorkReview:
    properties:
      body:
        allOf:
        - $ref: '#/definitions/storage.WorkReviewBody'
        description: BODY REVIEW
      created_date:
        type: string
      id:
        type: string
      language:
        type: string
      status:
        type: string
      updated_date:
        type: string
      work_id:
        type: string
    type: object
  storage.WorkReviewBody:
    properties:
      questionnaire:
        $ref: '#/definitions/storage.WorkReviewQuestionnaire'
      review:
        type: string
    type: object
  storage.WorkReviewQuestionnaire:
    properties:
      questions:
        additionalProperties:
          type: integer
        description: 0 - не согласен, 4 - согласен
        type: object
    type: object
info:
  contact: {}
  description: Specification of interaction with the application
  title: SOW library API
  version: "1.0"
paths:
...
  /become_validator:
    post:
      consumes:
      - application/json
      description: Become a validator
      parameters:
      - description: become validator
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/rest.BecomeValidatorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.AuthResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Become a validator
      tags:
      - Validators
...
  /submit_work_review:
    post:
      consumes:
      - application/json
      description: Submit work review by validator
      parameters:
      - description: work id
        in: path
        name: work_id
        required: true
        type: string
      - description: review status
        enum:
        - WORK_REVIEW_SUBMITTED
        - WORK_REVIEW_SKIPPED
        - WORK_REVIEW_REJECTED
        - WORK_REVIEW_DECLINED
        - WORK_REVIEW_ACCEPTED
        in: path
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Submit review
      tags:
      - Work review
...
  /update_review:
    post:
      consumes:
      - application/json
      description: Evaluate work by validator
      parameters:
      - description: work review
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/rest.WorkReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/storage.WorkReview'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Evaluate work
      tags:
      - Work review
  /update_validator_info:
    post:
      consumes:
      - application/json
      description: Update validator info
      parameters:
      - description: update validator info
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/rest.UpdateValidatorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.AuthResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Update validator
      tags:
      - Validators
...
  /validator_info/{web3_address}:
    get:
      consumes:
      - application/json
      description: Validator full info
      parameters:
      - description: validator web3 address
        in: path
        name: web3_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/storage.ValidatorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      summary: Validator info
      tags:
      - Validators
  /validator_info/upload_docs:
    post:
      consumes:
      - application/json
      description: Uploading documents confirming competencies of validator
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Upload validator documents
      tags:
      - Validators
...
  /work_review/{work_id}:
    get:
      consumes:
      - application/json
      description: Get all work reviews by work id
      parameters:
      - description: work id
        in: path
        name: work_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/storage.WorkReview'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Work reviews
      tags:
      - Work review
...
produces:
- application/json
schemes:
- http
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"